requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
colorama==0.4.6
colorlog==6.7.0
frozenlist==1.3.3
hikari==2.0.0.dev118
hikari-lightbulb==2.3.3
idna==3.4
import-expression==1.1.4
multidict==6.0.4
six==1.16.0
yarl==1.9.1
